[PATCH] zlib: Slim down zlib_deflate workspace when possible
Instead of always creating a huge (268K) deflate_workspace with the
maximum compression parameters (windowBits=15, memLevel=8), allow the
caller to obtain a smaller workspace by specifying smaller parameter
values -- via zlib_deflate_workspacesize2().

For example, when capturing oops and panic reports to a medium with
limited capacity, such as NVRAM, compression may be the only way to
capture the whole report.  In this case, a small workspace (24K works
fine) is a win, whether you allocate the workspace when you need it
(i.e., during an oops or panic) or at boot time.

I've verified that this patch works with all accepted values of
windowBits (positive and negative), memLevel, and compression level;
and also via the existing zlib_deflate_workspacesize() /
zlib_deflateInit() interface.

Signed-off-by: Jim Keniston <jkenisto@us.ibm.com>